pyproject.toml
[tool.black]
line-length = 119
exclude = ""

[tool.coverage.report]
# fail_under = 90
skip_covered = false
sort = "cover"

[tool.coverage.run]
command_line = "-m pytest"
omit = ["*pb2*"]

[tool.poetry]
name = "claco-pinochle"
version = "0.0.1"
description = "Pinochle Rummy (Rummy 500) Game"
authors = ["Christopher H. Laco <claco@chrislaco.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "pinochle", from = "src/python"}
]
include = ["poetry.lock", "README.md"]


[tool.poetry.dependencies]
python = "^3.10"
debugpy = "^1.6.4"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"
typer = "^0.7.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
coverage = "^6.5.0"
pytest-cov = "^4.0.0"
pytest-grpc = "^0.8.0"

[tool.poetry.scripts]
pinochle = "pinochle.cli:main"

[tool.pyright]
ignore = ["**/*_pb2*"]
reportWildcardImportFromLibrary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
